includes badges with get users
