+++ app/controllers/api/v1/auth_controller.rb
@@ -1,6 +1,8 @@
 module Api
   module V1
     class AuthController < ApiController
+      include Badgeable
+
       before_action :set_attrs
       skip_before_action :verify_authenticity_token
 
@@ -67,14 +69,6 @@ module Api
 
       def set_attrs
         @attrs = params.dig(:data, :attributes) || params.dig(:auth, :data, :attributes)
-      end
-
-      def random_badges_for(user)
-        user.user_badges.destroy_all
-        [:welcome, :observer, :addicted, :noob, :beginner].sample(rand(1..5)).each do |badge|
-          bdg = Badge.find_or_create_by(slug: badge, name: badge.to_s.titleize)
-          user.user_badges << UserBadge.new(badge_id: bdg.id, date: rand(1..2).days.ago)
-        end
       end
     end
   end
